Overhauled Knight's Collisions
Redefined how the knight interacts with the world. No new features added to the knight, but collisions are less buggy. This should be good for slope-less worlds. Also fixed a bug where 2 tiles in the level counted twice for collisions (which can lead to issues in this newly defined system).