Remade Collision
Remade collision to be somewhat nicer to look at, and hopefully understand. It works just like previously with improvement when in a small area and touching a wall.

The previous fix that prevented clipping is gone, and I seem to can't use the previous method. May try fixing it some other time.