Modified How Levels are Loaded
Levels are now loaded with the bottom left tile being the origin, and the (x,y) position works just like a standard cartesian graph on the 1st Quadrant. However, the definition of height goes downward. Also modified how the death floor work, which had catastrophic effect in the new system.