Improve NPC experience
Made the shop accessible only when idle at the shop, as well as when when enemies aren't nearby. Also activate checkpoint only when enemies aren't nearby. This allows the game to be a little bit user-friendly. It's still will punish you if enemies find their way to you regardless, but the act of using the shop makes more sense now.